+++ .gitignore
@@ -10,3 +10,5 @@ Image_Editor.egg-info/
 .pytest_cache
 */__pycache__/*
 *.pyc
+
+env/
