+++ image_editor/.env
@@ -1,0 +1,1 @@
+OP_DIR="output"
