+++ image_editor/utils/convert_image.py
@@ -3,6 +3,10 @@ from matplotlib import pyplot as plt
 from glob import glob
 import os
 from PIL import Image
+from dotenv import load_dotenv
+from pathlib import Path
+
+load_dotenv()
 
 
 def image_to_numpy_array(image_path: str) -> np.ndarray:
@@ -11,21 +15,23 @@ def image_to_numpy_array(image_path: str) -> np.ndarray:
 
 
 def numpy_array_to_image(image, image_path: str, file_extension=None) -> str:
-    if type(image) == np.ndarray:
-        img = Image.fromarray(image)
-    else:
-        img = Image.open(image)
-    new_file_name = create_new_image_file(image_path, file_extension)
-    img.save(new_file_name)
-    return new_file_name
-
+    img = Image.fromarray(image)
+    dirname, extension = create_new_image_file(image_path, file_extension)
+    
+    if not os.path.isdir(dirname):
+        os.makedirs( dirname)
+    
+    list_of_files = Path(dirname).glob(f'*.{extension}')
+    seq_no = max([0]+[int(Path(fn).stem.split('_')[1])  for fn in list_of_files if '_' in Path(fn).stem])
+    new_file_path =dirname + os.sep + f'OPIMG_{seq_no + 1}.{extension}'
+    img.save(new_file_path)
+    return new_file_path
 
 def create_new_image_file(image_path, file_extension):
     input_file_name = image_path.split(os.sep)[-1]
     if not file_extension:
         file_extension = input_file_name.split('.')[-1]
-    new_file_name = input_file_name.split('.')[0] + '_1.' + file_extension
-    return new_file_name
+    return os.getenv('OP_DIR') + os.sep + input_file_name.split('.')[0], file_extension
 
 
 def transparent_background(image):
@@ -38,7 +44,4 @@ def transparent_background(image):
         else:
             new_data.append(item)
     img.putdata(new_data)
-    img.save('temp.png', 'PNG')
-    while not glob('temp.png'):
-        pass
-    return 'temp.png'
+    return np.array(img)
