+++ requirements.txt
@@ -6,3 +6,5 @@ pytesseract==0.3.9
 numpy==1.22.3
 setuptools==62.1.0
 matplotlib==3.5.2
+python-dotenv==0.20.0
+pathlib==1.0.1
